requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
excel2json-3==0.1.6
idna==3.6
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
pyaes==1.6.1
pyasn1==0.5.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
TgCrypto==1.2.5
tzdata==2023.4
urllib3==2.1.0
xlrd==2.0.1